\Total:                   70.0 of 70.0\

+++ read-line.cc
@@ -222,7 +222,7 @@ char * read_line() {
         // down arrow
         // Erase old line
         // Print backspaces
-        if(history_index == history.size() && history_index == history.size() - 1)
+        if(history_index == (int) history.size() && history_index == (int)history.size() - 1)
           continue;
         int i = 0;
         for (i =0 ; i < line_length; i++) {
